Fixed bug in producteev filters.
If multiple users had the same name, then only
one of them would show up in the filters.

Also, if a user existed in multiple workspaces,
one of them would get picked randomly for the filter,
and all tasks assigned to that user but on a workspace
other than the randomly picked one would not show up.
Now all tasks by that user appear, regardless of
workspace.

Also, it's 9 lines shorter. :)