Fix for AST-363
Specifically, what this fixes is the case where a task has multiple alarms
in the future. Previously, only one of the alarms would run, but now they
are all scheduled as separate alarms and should all fire.

If there are other bugs not covered by this, please file another bug.