Added 'Filter By Tag' TaskListActivity context menu item
If the task for which the context menu has been created has no tags
this entry will be disabled (but visible).  For a single tag a new
TaskListActivity is started and with a filter based on that single
tag.  If multiple tags are present then a dialog is presented to allow
the user to select a single tag to filter by.